JAMES-1644 move GetMailboxesMethodTest in the right package
